refactor: extract verify helpers to eval_handlers.verification and make registry public
